just fixing stuff

+++ draft/fflas_reduce.cpp
@@ -6,8 +6,8 @@ extern "C" {
 
 uint32_t RowReduce_int32_t ( int32_t p, int32_t * A, uint32_t m, uint32_t n, uint32_t lda) ;
 uint32_t RowReduce_double  ( double p, double * A, uint32_t m, uint32_t n, uint32_t lda) ;
-void     Freduce_double(double p, double * A, uint32_t n);
-void     Finit_double  (double p, const double * A, double * B, uint32_t n);
+void     Freduce_double(double p, double * A, uint64_t n);
+void     Finit_double  (double p, const double * A, double * B, uint64_t n);
 
 // #include <cblas.h>
 
@@ -45,13 +45,13 @@ uint32_t RowReduce_double( double p, double * A, uint32_t m, uint32_t n, uint32_
 
 }
 
-void Freduce_double(double p, double * A, uint32_t n)
+void Freduce_double(double p, double * A, uint64_t n)
 {
 	FFPACK::Modular<double> F(p);
 	FFLAS::freduce(F,n,A,1);
 }
 
-void Finit_double(double p, const double * A, double * B, uint32_t n)
+void Finit_double(double p, const double * A, double * B, uint64_t n)
 {
 	FFPACK::Modular<double> F(p);
 	FFLAS::finit(F,n,A,1,B,1);
